Fix bug in stat filters count + refactoring (#432)
